requirements.txt
python-dateutil
requests
icalendar
pandas
gunicorn
frozen-flask